+++ Algorithms.h
@@ -5,106 +5,106 @@
 #include <algorithm>
 
 class Solution {
-	static bool is_linear(const std::vector<std::vector<unsigned int>>& _coefficients);
+	static bool is_linear(const std::vector<std::vector<unsigned long long>>& _coefficients);
 
 public:
-	const std::vector<std::vector<unsigned int>>& coefficients;
-	const unsigned int range;
+	const std::vector<std::vector<unsigned long long>>& coefficients;
+	const unsigned long long range;
 	const bool linear;
 
 	Solution(
-		const std::vector<std::vector<unsigned int>>& _coefficients,
-		const unsigned int _range
+		const std::vector<std::vector<unsigned long long>>& _coefficients,
+		const unsigned long long _range
 	);
 
-	std::pair<std::vector<unsigned int>, std::vector<unsigned int>> solve();
+	std::pair<std::vector<unsigned long long>, std::vector<unsigned long long>> solve();
 };
 
 // interface for our algorithms
 class Method {
 protected:
-	void reduce(std::vector<unsigned int>& equation);
+	void reduce(std::vector<unsigned long long>& equation);
 
 	const Solution* data_ptr;
 	void invalid_call() const throw(...);
 public:
 	Method(const Solution* _data);
 	
-	std::vector<unsigned int> answers;
-	std::vector<unsigned int> basis;
+	std::vector<unsigned long long> answers;
+	std::vector<unsigned long long> basis;
 
-	virtual std::pair<std::vector<unsigned int>, std::vector<unsigned int>> solve();
+	virtual std::pair<std::vector<unsigned long long>, std::vector<unsigned long long>> solve();
 
 };
 
 class Quine : public Method {
 	struct Value {
-		const unsigned int index;
+		const unsigned long long index;
 		const bool positive;
 
-		Value(unsigned int _index, bool _positive = true);
+		Value(unsigned long long _index, bool _positive = true);
 
-		int count(const unsigned int coef) const;
+		int count(const unsigned long long coef) const;
 
-		friend unsigned int operator+ (const Value& val, const unsigned int vec);
+		friend unsigned long long operator+ (const Value& val, const unsigned long long vec);
 
 	private:
-		unsigned int get_number() const;
+		unsigned long long get_number() const;
 	};
-	friend unsigned int operator+ (const Value& val, const unsigned int vec);
+	friend unsigned long long operator+ (const Value& val, const unsigned long long vec);
 
 	bool update_coefs(
-		std::vector<std::vector<unsigned int>>& coefs, 
+		std::vector<std::vector<unsigned long long>>& coefs, 
 		const Value& val
 	);
 
 	void set_value(
-		std::vector<std::vector<unsigned int>> coefs, 
-		const unsigned int vec, 
+		std::vector<std::vector<unsigned long long>> coefs, 
+		const unsigned long long vec, 
 		const Value& val
 	);
 
 	// find first unknown variable in current @coefs
-	unsigned int get_nearest_var(
-		const std::vector<std::vector<unsigned int>>& coefs
+	unsigned long long get_nearest_var(
+		const std::vector<std::vector<unsigned long long>>& coefs
 	);
 
 	// like dfs
 	void step(
-		const std::vector<std::vector<unsigned int>>& coefs, 
-		unsigned int vec = 0
+		const std::vector<std::vector<unsigned long long>>& coefs, 
+		unsigned long long vec = 0
 	);
 
 public:
 	// Note: zero coefficient means Positive value for equation
 	Quine(const Solution* solution);
 
-	std::pair<std::vector<unsigned int>, std::vector<unsigned int>> solve();
+	std::pair<std::vector<unsigned long long>, std::vector<unsigned long long>> solve();
 };
 
 class TSS : public Method {
-	std::vector<std::vector<unsigned int>> prepare_coefs(
-		std::vector<std::vector<unsigned int>> coefs, 
-		const unsigned int range) throw(...);
+	std::vector<std::vector<unsigned long long>> prepare_coefs(
+		std::vector<std::vector<unsigned long long>> coefs, 
+		const unsigned long long range) throw(...);
 
-	std::vector<unsigned int> get_start_vectors(const std::vector<unsigned int>& first, unsigned int range);
+	std::vector<unsigned long long> get_start_vectors(const std::vector<unsigned long long>& first, unsigned long long range);
 
-	bool multiply(const std::vector<unsigned int>& equation, const unsigned int value);
+	bool multiply(const std::vector<unsigned long long>& equation, const unsigned long long value);
 
-	unsigned int calcul_sum(const std::vector<unsigned int>& coefs, const unsigned int vec);
+	unsigned long long calcul_sum(const std::vector<unsigned long long>& coefs, const unsigned long long vec);
 
-	std::vector<unsigned int> add_equation(
-		const std::vector<unsigned int>& equation, 
-		const std::vector<unsigned int>& vectors
+	std::vector<unsigned long long> add_equation(
+		const std::vector<unsigned long long>& equation, 
+		const std::vector<unsigned long long>& vectors
 	);
 
 	// first -> individual, second -> basis
-	std::pair<std::vector<unsigned int>, std::vector<unsigned int>> separate_solutions(
-		const std::vector<unsigned int>& vec
+	std::pair<std::vector<unsigned long long>, std::vector<unsigned long long>> separate_solutions(
+		const std::vector<unsigned long long>& vec
 	);
 
 public:
 	TSS(const Solution* solution);
 
-	std::pair<std::vector<unsigned int>, std::vector<unsigned int>> solve();
+	std::pair<std::vector<unsigned long long>, std::vector<unsigned long long>> solve();
 };
