+++ templates.h
@@ -3,6 +3,6 @@
 #include <vector>
 #include <string>
 
-using u_i = unsigned int;
+using u_i = unsigned long long;
 template<typename T> using V = std::vector<T>;
 using S = std::string;
